Audio Manager and Volume Setting
Added the function of loop of BG Music and Continuously play the audio even new scene is loaded. Added the Volume Function where it save and load using playerPrefs implemented in Main Menu and Lesson. Fixed minor bug in the navbar UI where it doesn't render when scene is loaded.